adapt slab_new prototype for 64bit integers

+++ include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 17:46:51 by estettle          #+#    #+#             */
-/*   Updated: 2024/11/14 10:40:43 by estettle         ###   ########.fr       */
+/*   Updated: 2024/11/14 11:23:27 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void	ft_rrr(t_slab **stack1, t_slab **stack2);
 t_slab	**parse_input(int argc, char **argv);
 
 // Chained lists management function prototypes
-t_slab	*slab_new(int nb);
+t_slab	*slab_new(int64_t nb);
 t_slab	*slab_last(t_slab *lst);
 void	slab_add_back(t_slab **lst, t_slab *new);
 void	slab_add_front(t_slab **lst, t_slab *new);
